+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Paper and Purpur as server software
 - Metadata to keep information of what version of MCServer is used
 - State to keep information of a running MCServer to avoid conflict